+++ ddtrace/contrib/openai/__init__.py
@@ -253,10 +253,12 @@ required_modules = ["openai"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from . import patch as _patch
+        # Required to allow users to import from `ddtrace.contrib.openai.patch` directly
+        from . import patch as _  # noqa: F401, I001
 
-        patch = _patch.patch
-        unpatch = _patch.unpatch
-        get_version = _patch.get_version
+        # Expose public methods
+        from ..internal.openai.patch import patch
+        from ..internal.openai.patch import unpatch
+        from ..internal.openai.patch import get_version
 
         __all__ = ["patch", "unpatch", "get_version"]
