+++ ddtrace/contrib/psycopg/__init__.py
@@ -60,9 +60,19 @@ To configure the psycopg integration on an per-connection basis use the
     cursor = db.cursor()
     cursor.execute("select * from users where id = 1")
 """
-from .patch import get_version
-from .patch import get_versions
-from .patch import patch
+from ...internal.utils.importlib import require_modules
+
+
+required_modules = ["psycopg", "psycopg2"]
+with require_modules(required_modules) as missing_modules:
+    if not missing_modules:
+        # Required to allow users to import from `ddtrace.contrib.openai.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.psycopg.patch import get_version
+        from ..internal.psycopg.patch import get_versions
+        from ..internal.psycopg.patch import patch
 
 
 __all__ = ["patch", "get_version", "get_versions"]
