+++ ddtrace/contrib/psycopg/async_cursor.py
@@ -1,11 +1,15 @@
-from ddtrace.contrib import dbapi_async
-from ddtrace.contrib.psycopg.cursor import Psycopg3TracedCursor
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
+
+from ..internal.psycopg.async_cursor import *  # noqa: F401,F403
 
 
-class Psycopg3TracedAsyncCursor(Psycopg3TracedCursor, dbapi_async.TracedAsyncCursor):
-    def __init__(self, cursor, pin, cfg, *args, **kwargs):
-        super(Psycopg3TracedAsyncCursor, self).__init__(cursor, pin, cfg)
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-
-class Psycopg3FetchTracedAsyncCursor(Psycopg3TracedAsyncCursor, dbapi_async.FetchTracedAsyncCursor):
-    """Psycopg3FetchTracedAsyncCursor for psycopg"""
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
