+++ ddtrace/contrib/psycopg/cursor.py
@@ -1,28 +1,15 @@
-from ddtrace.contrib import dbapi
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
+
+from ..internal.psycopg.cursor import *  # noqa: F401,F403
 
 
-class Psycopg3TracedCursor(dbapi.TracedCursor):
-    """TracedCursor for psycopg instances"""
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-    def __init__(self, cursor, pin, cfg, *args, **kwargs):
-        super(Psycopg3TracedCursor, self).__init__(cursor, pin, cfg)
-
-    def _trace_method(self, method, name, resource, extra_tags, dbm_propagator, *args, **kwargs):
-        # treat Composable resource objects as strings
-        if resource.__class__.__name__ == "SQL" or resource.__class__.__name__ == "Composed":
-            resource = resource.as_string(self.__wrapped__)
-        return super(Psycopg3TracedCursor, self)._trace_method(
-            method, name, resource, extra_tags, dbm_propagator, *args, **kwargs
-        )
-
-
-class Psycopg3FetchTracedCursor(Psycopg3TracedCursor, dbapi.FetchTracedCursor):
-    """Psycopg3FetchTracedCursor for psycopg"""
-
-
-class Psycopg2TracedCursor(Psycopg3TracedCursor):
-    """TracedCursor for psycopg2"""
-
-
-class Psycopg2FetchTracedCursor(Psycopg3FetchTracedCursor):
-    """FetchTracedCursor for psycopg2"""
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
