+++ ddtrace/contrib/pylibmc/__init__.py
@@ -26,8 +26,12 @@ required_modules = ["pylibmc"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .client import TracedClient
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.pylibmc.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.pylibmc.client import TracedClient
+        from ..internal.pylibmc.patch import get_version
+        from ..internal.pylibmc.patch import patch
 
         __all__ = ["TracedClient", "patch", "get_version"]
