+++ ddtrace/contrib/pymemcache/__init__.py
@@ -37,8 +37,12 @@ required_modules = ["pymemcache"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.pymemcache.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.pymemcache.patch import get_version
+        from ..internal.pymemcache.patch import patch
+        from ..internal.pymemcache.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
