+++ ddtrace/contrib/pymongo/__init__.py
@@ -42,7 +42,11 @@ required_modules = ["pymongo"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.pymongo.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.pymongo.patch import get_version
+        from ..internal.pymongo.patch import patch
 
         __all__ = ["patch", "get_version"]
