+++ ddtrace/contrib/pymysql/__init__.py
@@ -62,7 +62,11 @@ required_modules = ["pymysql"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.pymysql.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.pymysql.patch import get_version
+        from ..internal.pymysql.patch import patch
 
         __all__ = ["patch", "get_version"]
