+++ tests/contrib/openai/test_openai_v0.py
@@ -10,7 +10,7 @@ import pytest
 
 import ddtrace
 from ddtrace import patch
-from ddtrace.contrib.openai.utils import _est_tokens
+from ddtrace.contrib.internal.openai.utils import _est_tokens
 from ddtrace.internal.utils.version import parse_version
 from tests.contrib.openai.utils import chat_completion_custom_functions
 from tests.contrib.openai.utils import chat_completion_input_description
@@ -1264,7 +1264,7 @@ def test_span_finish_on_stream_error(openai, openai_vcr, snapshot_tracer):
 
 def test_completion_stream(openai, openai_vcr, mock_metrics, mock_tracer):
     with openai_vcr.use_cassette("completion_streamed.yaml"):
-        with mock.patch("ddtrace.contrib.openai.utils.encoding_for_model", create=True) as mock_encoding:
+        with mock.patch("ddtrace.contrib.internal.openai.utils.encoding_for_model", create=True) as mock_encoding:
             mock_encoding.return_value.encode.side_effect = lambda x: [1, 2]
             expected_completion = '! ... A page layouts page drawer? ... Interesting. The "Tools" is'
             resp = openai.Completion.create(model="ada", prompt="Hello world", stream=True)
@@ -1305,7 +1305,7 @@ def test_completion_stream(openai, openai_vcr, mock_metrics, mock_tracer):
 @pytest.mark.asyncio
 async def test_completion_async_stream(openai, openai_vcr, mock_metrics, mock_tracer):
     with openai_vcr.use_cassette("completion_async_streamed.yaml"):
-        with mock.patch("ddtrace.contrib.openai.utils.encoding_for_model", create=True) as mock_encoding:
+        with mock.patch("ddtrace.contrib.internal.openai.utils.encoding_for_model", create=True) as mock_encoding:
             mock_encoding.return_value.encode.side_effect = lambda x: [1, 2]
             expected_completion = "\" and just start creating stuff. Don't expect it to draw like this."
             resp = await openai.Completion.acreate(model="ada", prompt="Hello world", stream=True)
@@ -1345,7 +1345,7 @@ def test_chat_completion_stream(openai, openai_vcr, mock_metrics, snapshot_trace
         pytest.skip("ChatCompletion not supported for this version of openai")
 
     with openai_vcr.use_cassette("chat_completion_streamed.yaml"):
-        with mock.patch("ddtrace.contrib.openai.utils.encoding_for_model", create=True) as mock_encoding:
+        with mock.patch("ddtrace.contrib.internal.openai.utils.encoding_for_model", create=True) as mock_encoding:
             mock_encoding.return_value.encode.side_effect = lambda x: [1, 2, 3, 4, 5, 6, 7, 8]
             expected_completion = "The Los Angeles Dodgers won the World Series in 2020."
             resp = openai.ChatCompletion.create(
@@ -1395,7 +1395,7 @@ async def test_chat_completion_async_stream(openai, openai_vcr, mock_metrics, sn
     if not hasattr(openai, "ChatCompletion"):
         pytest.skip("ChatCompletion not supported for this version of openai")
     with openai_vcr.use_cassette("chat_completion_streamed_async.yaml"):
-        with mock.patch("ddtrace.contrib.openai.utils.encoding_for_model", create=True) as mock_encoding:
+        with mock.patch("ddtrace.contrib.internal.openai.utils.encoding_for_model", create=True) as mock_encoding:
             mock_encoding.return_value.encode.side_effect = lambda x: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
             expected_completion = "As an AI language model, I do not have access to real-time information but as of the 2021 season, the captain of the Toronto Maple Leafs is John Tavares."  # noqa: E501
             resp = await openai.ChatCompletion.acreate(
